backend/requirements.txt
psycopg2>=2.8.6,<2.8.7
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
six==1.15.0
PyJWT==v1.7.1
django-cors-headers
django-extensions
